requirements.txt
Pyrogram
Tgcrypto
python-dotenv
flask
gunicorn
psutil
# uvloop
motor